REFACT: Move well network mesh and intersection detection methods inside the class
